integrators/other/email-client/v1.0.0/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
aiokafka==0.12.0
aiosqlite==0.20.0
prometheus-client==0.21.1
pyyaml==6.0.2
